+++ src/CodeFixes/CSharp/CodeFixIdentifiers.Generated.cs
@@ -123,5 +123,6 @@ namespace Roslynator.CSharp
         public const string RemoveAnnotationForNullableReferenceTypes = CodeFixIdentifier.CodeFixIdPrefix + "0118";
         public const string UseNullForgivingOperator = CodeFixIdentifier.CodeFixIdPrefix + "0119";
         public const string AddAsyncModifier = CodeFixIdentifier.CodeFixIdPrefix + "0120";
+        public const string UseNullPropagationOperator = CodeFixIdentifier.CodeFixIdPrefix + "0121";
     }
 }