+++ ChangeLog.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+* Add code fix for CS8602, CS8604
 * Fix code fix for CS0225
 
 ### 4.0.3 (2022-01-29)